btrfs: split-brain case for scanned changed device without INCOMPAT_METADATA_UUID
This patch adds the case for scanned not changing device without
INCOMPAT_METADATA_UUID.
For this situation, the origin code only handles the case
the devices already pulled into disk with INCOMPAT_METADATA_UUID set.
There is an another case that the successful changed devices synced
without INCOMPAT_METADATA_UUID. And the scanned device is the exactly
changed one.

So we should check if there is any changing fs_devices with
INCOMPAT_METADATA_UUID.

Signed-off-by: Su Yue <Damenly_Su@gmx.com>